requirements.txt
crewai
crewai_tools
langchain-community
langchain_groq
python-dotenv
pydantic
